scale LISI score using a csv file as input
The LISI score is the median over all observed scores in a sample, where the values range from 1 to some maximum value. A scaling that uses the highest median value in one scenario mean that we assign the highest possible score to the best performing method. Thus, the reported LISI score does not show, if the method performed well overall. We tweak this a little bit by using a minimal interval of length 1 for scaling.